some changes to allow compilation with sun compiler

+++ applications/incompressible_fluid_application/custom_processes/find_intersections_process.h
@@ -401,7 +401,7 @@ namespace Kratos
 				if (n_points_in_radius>=1)
 				{										
 					//find to which edge (condition) of the origin mesh this point belongs
-					for(PointIterator i=res.begin(); i!=res.begin() + n_points_in_radius ; i++)
+				  for(PointIterator i=res.begin(); i!=res.begin() + static_cast<int>(n_points_in_radius) ; i++)
 						{
 						WeakPointerVector< Condition >& neighb_conds = (*i)->GetValue(NEIGHBOUR_CONDITIONS);
